Adding local.py
This is basically the same as the local-rdd branch, but I’ve further
subclassed LocalRDD with LocalNumpyRDD which uses a numpy array as the
implementation for the store.  This lets us use numpy indexing for
groupbykey, which was is the largest bottleneck for local performance.
All the other methods are inherited.

I am not sure how consistent this implementation is, that is, if you
create a LocalNumpyRDD with different types of objects, or use the
LocalNumpyRDD that results from groupbykey or reducebykey, it might
have a different form than one initialized by a list of tuples or
ndarry.  Probably we need to try to use this in some tests.

See these associated benchmark ipynbs:
http://nbviewer.ipython.org/330336d5180dedfbde3d (dist v. list)
http://nbviewer.ipython.org/797296cc54410dd39d3e (dist v. ndarray)